# chore: Flatten Uint64StatCollection