This commit fixes a couple of small issues.
Firstly, in Go maps are reference types (like pointers) so
using *map[string]any for example isn't required.
Secondly, this removes the structure that backs
Uint64StatCollection and replaces it with a map.
Doing so make the code much easier to read (IMHO).

Finally, operations in Uint64Stat have been made
atomic to ensure correct operation if methods are
called from different threads. A test has been added
to verify this behaviour - this test fails when
run against the main branch.

There are some additional small fixes to address lint issues
that appear to have eluded our CI. Firstly the `dupl` threshold
has been moved to 150 to stop it complaining about Gingko-style
tests (which are very similar). Secondly the YAML for the `shadow`
has been fixed and issues with variable shadowing also fixed.

Below are the benchmark results vs. the implementaton in main.
``` console
$ benchstat ../main.txt ../uintflat.txt
goos: linux
goarch: amd64
pkg: github.com/sustainable-computing-io/kepler/pkg/collector/stats/types
cpu: AMD Ryzen 5 5600G with Radeon Graphics         
                                                      │ ../main.txt  │            ../uintflat.txt            │
                                                      │    sec/op    │    sec/op      vs base                │
UInt64StatCollectionSetAggrStatMissed-12                480.0n ± 20%    472.8n ± 17%        ~ (p=0.739 n=10)
UInt64StatCollectionSetAggrStatBothRandom-12            478.7n ±  2%    458.1n ±  3%   -4.28% (p=0.002 n=10)
UInt64StatCollectionSetAggrStatCachedRandomNumber-12    13.15n ±  1%    13.90n ±  0%   +5.71% (p=0.000 n=10)
UInt64StatCollectionSetAggrStatCached-12                5.301n ±  1%    5.609n ±  3%   +5.81% (p=0.000 n=10)
UInt64StatCollectionAddDeltaStatMissed-12               485.6n ±  3%    467.1n ±  8%        ~ (p=0.197 n=10)
UInt64StatCollectionAddDeltaStatBothRandom-12           478.2n ±  5%    400.8n ± 18%  -16.20% (p=0.009 n=10)
UInt64StatCollectionAddDeltaStatCachedRandomNumber-12   9.995n ±  1%   10.535n ±  1%   +5.40% (p=0.000 n=10)
UInt64StatCollectionAddDeltaStatCached-12               5.540n ±  1%    6.543n ±  1%  +18.09% (p=0.000 n=10)
UInt64StatCollectionSetDeltaStatMissed-12               487.8n ±  4%    473.0n ±  6%        ~ (p=0.165 n=10)
UInt64StatCollectionSetDeltaStatBothRandom-12           475.4n ±  3%    461.5n ±  3%   -2.91% (p=0.035 n=10)
UInt64StatCollectionSetDeltaStatCachedRandomNumber-12   10.26n ±  1%    10.79n ±  0%   +5.11% (p=0.000 n=10)
UInt64StatCollectionSetDeltaStatCached-12               5.785n ±  0%    6.792n ±  1%  +17.41% (p=0.000 n=10)
geomean                                                 61.34n          62.40n         +1.74%

                                                      │  ../main.txt  │           ../uintflat.txt            │
                                                      │     B/op      │    B/op      vs base                 │
UInt64StatCollectionSetAggrStatMissed-12                163.5 ± 33%     164.5 ± 33%       ~ (p=0.381 n=10)
UInt64StatCollectionSetAggrStatBothRandom-12            168.0 ±  1%     172.0 ±  2%  +2.38% (p=0.000 n=10)
UInt64StatCollectionSetAggrStatCachedRandomNumber-12    0.000 ±  0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionSetAggrStatCached-12                0.000 ±  0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionAddDeltaStatMissed-12               163.0 ±  2%     167.5 ±  4%  +2.76% (p=0.008 n=10)
UInt64StatCollectionAddDeltaStatBothRandom-12           168.0 ±  2%     110.0 ± 56%       ~ (p=0.493 n=10)
UInt64StatCollectionAddDeltaStatCachedRandomNumber-12   0.000 ±  0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionAddDeltaStatCached-12               0.000 ±  0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionSetDeltaStatMissed-12               164.0 ±  3%     167.0 ±  3%       ~ (p=0.098 n=10)
UInt64StatCollectionSetDeltaStatBothRandom-12           170.0 ±  1%     173.0 ±  1%  +1.76% (p=0.017 n=10)
UInt64StatCollectionSetDeltaStatCachedRandomNumber-12   0.000 ±  0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionSetDeltaStatCached-12               0.000 ±  0%     0.000 ±  0%       ~ (p=1.000 n=10) ¹
geomean                                                             ²                -2.72%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                                      │ ../main.txt  │           ../uintflat.txt           │
                                                      │  allocs/op   │ allocs/op   vs base                 │
UInt64StatCollectionSetAggrStatMissed-12                2.000 ± 0%     2.000 ± 0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionSetAggrStatBothRandom-12            2.000 ± 0%     2.000 ± 0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionSetAggrStatCachedRandomNumber-12    0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionSetAggrStatCached-12                0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionAddDeltaStatMissed-12               2.000 ± 0%     2.000 ± 0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionAddDeltaStatBothRandom-12           2.000 ± 0%     2.000 ± 0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionAddDeltaStatCachedRandomNumber-12   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionAddDeltaStatCached-12               0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionSetDeltaStatMissed-12               2.000 ± 0%     2.000 ± 0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionSetDeltaStatBothRandom-12           2.000 ± 0%     2.000 ± 0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionSetDeltaStatCachedRandomNumber-12   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
UInt64StatCollectionSetDeltaStatCached-12               0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                            ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

Note there is no significant performance difference between these 2 implementations.